+++ src/context/ProjectContext.jsx
@@ -17,17 +17,17 @@ const defaultProject = {
     {
       type: 'hh',
       steps: ['C3', 'C3', 'C3', 'C3', 'C3', 'C3', 'C3', 'C3','C3', 'C3', 'C3', 'C3', 'C3', 'C3', 'C3', 'C3'],
-      volume: 1
+      volume: -40
     },
     {
       type: 'snare',
       steps: [null, null, 'C3', null, null, null, 'C3', null, null, null, 'C3', null, null, null, 'C3', null,],
-      volume: 1
+      volume: -40
     },
     {
       type: 'kick',
       steps: ['C3', null, null, null, 'C3', null, null, null, 'C3', null, null, null, 'C3', null, null, null,],
-      volume: 1
+      volume: -40
     },
   ]
 };
